New pipeline for ExercisePlan. ExercisePlan 1-m Template 1-m ExerciseSlot[exerciseId : Integer] 1-m Week 1-m ExerciseSet
