+++ src/main/java/ucmo/workoutapp/services/ExerciseSetService.java
@@ -4,8 +4,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import ucmo.workoutapp.entities.ExerciseSet;
 import ucmo.workoutapp.entities.ExerciseSlot;
+import ucmo.workoutapp.entities.Week;
 import ucmo.workoutapp.repositories.ExerciseSetRepository;
 import ucmo.workoutapp.repositories.ExerciseSlotRepository;
+import ucmo.workoutapp.repositories.WeekRepository;
 
 @Service
 public class ExerciseSetService {
@@ -13,23 +15,25 @@ public class ExerciseSetService {
     private ExerciseSetRepository exerciseSetRepository;
 
     @Autowired
-    private ExerciseSlotRepository exerciseSlotRepository;
+    private WeekRepository weekRepository;
 
-    public ExerciseSet createExerciseSetForExerciseSlot(ExerciseSet exerciseSet, Long exerciseSlotId, String username){
-        ExerciseSlot exerciseSlot = exerciseSlotRepository.getById(exerciseSlotId);
-        exerciseSet.setExerciseSlot(exerciseSlot);
+    public ExerciseSet createExerciseSetForWeek(ExerciseSet exerciseSet, Long weekId, String username){
+        Week week = weekRepository.getById(weekId);
+        exerciseSet.setWeek(week);
 
-        // Get the number of sets belonging to the exercise slot
-        Integer setNumber = exerciseSlot.getSets().size();
+        // Get the number of sets belonging to the week
+        Integer exerciseSetNumber = week.getExerciseSets().size();
 
         // Increment to represent which set this is locally (separate from the ID. It will not be unique in the db).
         // e.g. if there are 0, this will now be set 1. If there are 4, this will now be set 5.
-        setNumber++;
+        exerciseSetNumber++;
 
-        // Set the set number to this new number
-        exerciseSet.setSetNumber(setNumber);
+        // Set the ExerciseSet number to this new number
+        exerciseSet.setExerciseSetNumber(exerciseSetNumber);
 
         return exerciseSetRepository.save(exerciseSet);
+
+        //**************** There's probably a better way to name this ****************************//
     }
 
     public ExerciseSet getExerciseSetById(Long exerciseSetId, String username){
