+++ src/main/java/ucmo/workoutapp/services/WeekService.java
@@ -2,6 +2,7 @@ package ucmo.workoutapp.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import ucmo.workoutapp.entities.ExerciseSlot;
 import ucmo.workoutapp.entities.Template;
 import ucmo.workoutapp.entities.Week;
 import ucmo.workoutapp.repositories.*;
@@ -9,23 +10,17 @@ import ucmo.workoutapp.repositories.*;
 @Service
 public class WeekService {
     @Autowired
-    private ExercisePlanRepository exercisePlanRepository;
-
-    @Autowired
-    private TemplateRepository templateRepository;
+    private ExerciseSlotRepository exerciseSlotRepository;
 
     @Autowired
     private WeekRepository weekRepository;
 
     @Autowired
-    private UserRepository userRepository;
-
-    @Autowired
     private ClientRepository clientRepository;
 
-    public Week createWeekForTemplate(Week week, Long templateId, String username){
-        Template template = templateRepository.getById(templateId);
-        week.setTemplate(template);
+    public Week createWeekForExerciseSlot(Week week, Long exerciseSlotId, String username){
+        ExerciseSlot exerciseSlot = exerciseSlotRepository.getById(exerciseSlotId);
+        week.setExerciseSlot(exerciseSlot);
         week.setName(week.getName());
 
         return weekRepository.save(week);
